Move utils over to luigi util
